remove debug info + move filesystem code
